requirements.txt
## For formatting
enlighten
blessed
rich

## For google services
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

## Images
pillow